chore: upgrade Universal demo to .NET Core 2.2
