+++ examples-standalone/angular-universal/aspnet_core_sample_application/ClientApp/.gitignore
@@ -31,12 +31,9 @@
 /coverage
 /libpeerconnection.log
 npm-debug.log
+yarn-error.log
 testem.log
 /typings
-
-# e2e
-/e2e/*.js
-/e2e/*.map
 
 # System Files
 .DS_Store
