+++ examples-standalone/angular-universal/aspnet_core_sample_application/ClientApp/e2e/src/app.e2e-spec.ts
@@ -1,0 +1,14 @@
+import { AppPage } from './app.po';
+
+describe('App', () => {
+  let page: AppPage;
+
+  beforeEach(() => {
+    page = new AppPage();
+  });
+
+  it('should display welcome message', () => {
+    page.navigateTo();
+    expect(page.getMainHeading()).toEqual('Hello, world!');
+  });
+});
