+++ examples-standalone/angular-universal/aspnet_core_sample_application/ClientApp/e2e/src/app.po.ts
@@ -1,0 +1,11 @@
+import { browser, by, element } from 'protractor';
+
+export class AppPage {
+  navigateTo() {
+    return browser.get('/');
+  }
+
+  getMainHeading() {
+    return element(by.css('app-root h1')).getText();
+  }
+}
