+++ examples-standalone/angular-universal/aspnet_core_sample_application/ClientApp/package.json
@@ -1,64 +1,71 @@
 {
   "name": "dotnet_angular",
   "version": "0.0.0",
-  "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --extract-css",
-    "build": "ng build --extract-css",
-    "build:ssr": "npm run build -- --app=ssr --output-hashing=media",
+    "start": "ng serve",
+    "build": "ng build",
+    "build:prod": "ng build --prod",
+    "build:ssr": "ng run dotnet_angular:server:production",
+    "build:full:prod": "npm run build:prod && npm run build:ssr",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/platform-server": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@nguniversal/module-map-ngfactory-loader": "^5.0.0-beta.5",
-    "@progress/kendo-angular-buttons": "^3.0.1",
-    "@progress/kendo-angular-dateinputs": "^2.0.0",
-    "@progress/kendo-angular-dropdowns": "^2.0.0",
-    "@progress/kendo-angular-excel-export": "^1.0.5",
-    "@progress/kendo-angular-grid": "^2.0.1",
-    "@progress/kendo-angular-inputs": "^2.0.0",
-    "@progress/kendo-angular-intl": "^1.3.0",
-    "@progress/kendo-angular-l10n": "^1.0.5",
-    "@progress/kendo-data-query": "^1.1.2",
-    "@progress/kendo-drawing": "^1.4.1",
-    "@progress/kendo-theme-default": "^2.47.2",
+    "@angular-devkit/build-angular": "^0.11.4",
+    "@angular/animations": "^7.2.3",
+    "@angular/cli": "^7.3.0",
+    "@angular/common": "^7.2.3",
+    "@angular/compiler": "^7.2.3",
+    "@angular/compiler-cli": "^7.2.3",
+    "@angular/core": "^7.2.3",
+    "@angular/forms": "^7.2.3",
+    "@angular/http": "^7.2.3",
+    "@angular/platform-browser": "^7.2.3",
+    "@angular/platform-browser-dynamic": "^7.2.3",
+    "@angular/platform-server": "^7.2.3",
+    "@angular/router": "^7.2.3",
+    "@nguniversal/module-map-ngfactory-loader": "^7.0.2",
+    "@progress/kendo-angular-buttons": "^4.0.0",
+    "@progress/kendo-angular-dateinputs": "2 - 3",
+    "@progress/kendo-angular-dropdowns": "2 - 3",
+    "@progress/kendo-angular-excel-export": "1 - 2",
+    "@progress/kendo-angular-grid": "^3.14.1",
+    "@progress/kendo-angular-inputs": "2 - 4",
+    "@progress/kendo-angular-intl": "^1.0.0",
+    "@progress/kendo-angular-l10n": "^1.1.0",
+    "@progress/kendo-angular-popup": "^2.0.0",
+    "@progress/kendo-data-query": "^1.0.0",
+    "@progress/kendo-drawing": "^1.0.0",
+    "@progress/kendo-theme-default": "^3.5.1",
     "aspnet-prerendering": "^3.0.1",
-    "bootstrap": "^3.3.7",
-    "core-js": "^2.4.1",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.14"
+    "babel-cli": "^6.26.0",
+    "babel-preset-es2015": "^6.24.1",
+    "bootstrap": "^4.2.1",
+    "classlist.js": "^1.1.20150312",
+    "core-js": "^2.6.1",
+    "rxjs": "^6.4.0",
+    "rxjs-compat": "^6.0.0",
+    "webpack-cli": "^3.1.2",
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.8",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@types/node": "^10.12.18",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~3.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "karma-phantomjs-launcher": "^1.0.2",
+    "lodash": "^4.16.2",
+    "phantomjs-prebuilt": "^2.1.7",
+    "protractor": "~5.4.0",
+    "tslint": "~5.11.0",
+    "ts-node": "~7.0.0",
+    "typescript": "~3.1.1"
   }
 }
