+++ examples-standalone/angular-universal/aspnet_core_sample_application/ClientApp/src/app/app.component.html
@@ -1,10 +1,6 @@
-<div class='container-fluid'>
-  <div class='row'>
-    <div class='col-sm-3'>
-      <app-nav-menu></app-nav-menu>
-    </div>
-    <div class='col-sm-9 body-content'>
-      <router-outlet></router-outlet>
-    </div>
+<body>
+  <app-nav-menu></app-nav-menu>
+  <div class="container">
+    <router-outlet></router-outlet>
   </div>
-</div>
+</body>
