+++ examples-standalone/angular-universal/aspnet_core_sample_application/ClientApp/src/app/app.module.ts
@@ -9,14 +9,9 @@ import { NavMenuComponent } from './nav-menu/nav-menu.component';
 import { HomeComponent } from './home/home.component';
 import { CounterComponent } from './counter/counter.component';
 import { FetchDataComponent } from './fetch-data/fetch-data.component';
-
-// Import the Animations module
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-
-// Import the ButtonsModule
-import { ButtonsModule } from '@progress/kendo-angular-buttons';
-
 import { GridModule } from '@progress/kendo-angular-grid';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+
 
 
 @NgModule({
@@ -29,16 +24,15 @@ import { GridModule } from '@progress/kendo-angular-grid';
   ],
   imports: [
     BrowserModule.withServerTransition({ appId: 'ng-cli-universal' }),
-    ButtonsModule,
-    GridModule,
-    BrowserAnimationsModule,
     HttpClientModule,
     FormsModule,
     RouterModule.forRoot([
       { path: '', component: HomeComponent, pathMatch: 'full' },
       { path: 'counter', component: CounterComponent },
       { path: 'fetch-data', component: FetchDataComponent },
-    ])
+    ]),
+    GridModule,
+    NoopAnimationsModule
   ],
   providers: [],
   bootstrap: [AppComponent]
