+++ examples-standalone/angular-universal/aspnet_core_sample_application/ClientApp/src/app/counter/counter.component.html
@@ -4,4 +4,4 @@
 
 <p>Current count: <strong>{{ currentCount }}</strong></p>
 
-<button (click)="incrementCounter()">Increment</button>
+<button class="btn btn-primary" (click)="incrementCounter()">Increment</button>
