+++ examples-standalone/angular-universal/aspnet_core_sample_application/ClientApp/src/app/home/home.component.html
@@ -12,5 +12,3 @@
   <li><strong>Efficient production builds</strong>. In production mode, development-time features are disabled, and your <code>dotnet publish</code> configuration automatically invokes <code>ng build</code> to produce minified, ahead-of-time compiled JavaScript files.</li>
 </ul>
 <p>The <code>ClientApp</code> subdirectory is a standard Angular CLI application. If you open a command prompt in that directory, you can run any <code>ng</code> command (e.g., <code>ng test</code>), or use <code>npm</code> to install extra packages into it.</p>
-<p>If you want to enable server-side prerendering, see the steps in <code>Startup.cs</code>.</p>
-<h3>A fully functional Kendo UI for Angular Grid is present in the <strong>Fetch data</strong> page.</h3>
