+++ examples-standalone/angular-universal/aspnet_core_sample_application/ClientApp/src/app/nav-menu/nav-menu.component.ts
