+++ examples-standalone/angular-universal/aspnet_core_sample_application/ClientApp/src/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
   <meta charset="utf-8">
-  <title>dotnet_angular</title>
+  <title>.NET Core - Angular Universal</title>
   <base href="/">
 
   <meta name="viewport" content="width=device-width, initial-scale=1">
