+++ examples-standalone/angular-universal/aspnet_core_sample_application/ClientApp/src/main.ts
