+++ examples-standalone/angular-universal/aspnet_core_sample_application/ClientApp/src/styles.scss
@@ -1,4 +1,1 @@
 /* You can add global styles to this file, and also import other style files */
-
-@import "~@progress/kendo-theme-default/scss/all";
-
