+++ examples-standalone/angular-universal/aspnet_core_sample_application/ClientApp/src/tsconfig.server.json
@@ -1,11 +1,13 @@
 {
-    "extends": "../tsconfig.json",
-    "compilerOptions": {
-      "baseUrl": "./",
-      "module": "commonjs"
-    },
-    "angularCompilerOptions": {
-      "entryModule": "app/app.server.module#AppServerModule"
-    }
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "module": "commonjs"
+  },
+  "angularCompilerOptions": {
+    "entryModule": "app/app.server.module#AppServerModule"
+  },
+  "exclude": [
+    "src/test.ts",
+    "**/*.spec.ts"
+  ]
 }
-
