+++ examples-standalone/angular-universal/aspnet_core_sample_application/Controllers/SampleDataController.cs
@@ -20,7 +20,9 @@ namespace dotnet_angular.Controllers
             var rng = new Random();
             return Enumerable.Range(1, 5).Select(index => new WeatherForecast
             {
-                DateFormatted = DateTime.Now.AddDays(index).ToString("d"),
+                // Use the "Round-trip Format Specifier" for dates
+                // https://docs.microsoft.com/en-us/dotnet/standard/base-types/standard-date-and-time-format-strings#the-round-trip-o-o-format-specifier
+                DateFormatted = DateTime.Now.AddDays(index).ToString("o"),
                 TemperatureC = rng.Next(-20, 55),
                 Summary = Summaries[rng.Next(Summaries.Length)]
             });
