+++ examples-standalone/angular-universal/aspnet_core_sample_application/Pages/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+@using dotnet_angular
+@namespace dotnet_angular.Pages
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
