+++ examples-standalone/angular-universal/aspnet_core_sample_application/Program.cs
