+++ examples-standalone/angular-universal/aspnet_core_sample_application/appsettings.Development.json
@@ -1,4 +1,4 @@
-﻿{
+{
   "Logging": {
     "LogLevel": {
       "Default": "Debug",
