+++ examples-standalone/angular-universal/aspnet_core_sample_application/appsettings.json
@@ -3,5 +3,6 @@
     "LogLevel": {
       "Default": "Warning"
     }
-  }
+  },
+  "AllowedHosts": "*"
 }
